+++ Tests/Command/DebugCommandTest.php
@@ -225,6 +225,87 @@ class DebugCommandTest extends TestCase
         ];
     }
 
+    public function testExecuteWithSortAndAllListsNullNextRunFirst()
+    {
+        $schedule = new Schedule();
+        $schedule
+            ->add(RecurringMessage::every('2 minutes', new \stdClass()))
+            ->add(RecurringMessage::trigger(new CallbackTrigger(static fn () => null, 'terminated'), new \stdClass()))
+            ->add(RecurringMessage::every('1 minute', new \stdClass()))
+        ;
+
+        $schedules = $this->createMock(ServiceProviderInterface::class);
+        $schedules
+            ->expects($this->once())
+            ->method('getProvidedServices')
+            ->willReturn(['schedule_name' => $schedule])
+        ;
+        $schedules
+            ->expects($this->once())
+            ->method('get')
+            ->willReturn($schedule)
+        ;
+
+        $command = new DebugCommand($schedules);
+        $tester = new CommandTester($command);
+
+        $tester->execute(['--sort' => true, '--all' => true], ['decorated' => false]);
+
+        $this->assertMatchesRegularExpression("/\n".
+            "Scheduler\n".
+            "=========\n".
+            "\n".
+            "schedule_name\n".
+            "-------------\n".
+            "\n".
+            " ----------------- ---------- --------------------------------- \n".
+            "  Trigger           Provider   Next Run                         \n".
+            " ----------------- ---------- --------------------------------- \n".
+            "  terminated        stdClass   -                                \n".
+            "  every 1 minute    stdClass   \w{3}, \d{1,2} \w{3} \d{4} \d{2}:\d{2}:\d{2} (\+|-)\d{4}  \n".
+            "  every 2 minutes   stdClass   \w{3}, \d{1,2} \w{3} \d{4} \d{2}:\d{2}:\d{2} (\+|-)\d{4}  \n".
+            " ----------------- ---------- --------------------------------- \n".
+            "\n/", $tester->getDisplay(true));
+    }
+
+    public function testExecuteWithSortPreservesInsertionOrderOnTies()
+    {
+        $schedule = new Schedule();
+        $schedule
+            ->add(RecurringMessage::trigger(new CallbackTrigger(static fn (\DateTimeImmutable $run) => $run->modify('+1 minute'), 'first'), new \stdClass()))
+            ->add(RecurringMessage::trigger(new CallbackTrigger(static fn (\DateTimeImmutable $run) => $run->modify('+1 minute'), 'second'), new \stdClass()))
+            ->add(RecurringMessage::trigger(new CallbackTrigger(static fn (\DateTimeImmutable $run) => $run->modify('+1 minute'), 'third'), new \stdClass()))
+        ;
+
+        $schedules = $this->createMock(ServiceProviderInterface::class);
+        $schedules
+            ->expects($this->once())
+            ->method('getProvidedServices')
+            ->willReturn(['schedule_name' => $schedule])
+        ;
+        $schedules
+            ->expects($this->once())
+            ->method('get')
+            ->willReturn($schedule)
+        ;
+
+        $command = new DebugCommand($schedules);
+        $tester = new CommandTester($command);
+
+        $tester->execute(['--sort' => true], ['decorated' => false]);
+
+        $display = $tester->getDisplay(true);
+        $firstPos = strpos($display, 'first');
+        $secondPos = strpos($display, 'second');
+        $thirdPos = strpos($display, 'third');
+
+        $this->assertNotFalse($firstPos);
+        $this->assertNotFalse($secondPos);
+        $this->assertNotFalse($thirdPos);
+        $this->assertLessThan($secondPos, $firstPos);
+        $this->assertLessThan($thirdPos, $secondPos);
+    }
+
     public function testExecuteWithStatefulScheduleUsesStoredCheckpoint()
     {
         $cache = new ArrayAdapter();